Only collect unique batch file paths
If Gem.path contains only a single entry, doing unshift with Gem.bindir
will result in a duplicated list of paths.

Having duplicate list of paths will result in attempt to copy the
generated executable twice and the double removal of batch file, which
will result on a exception the second time.