Generate executable only once
The idea with these changes (experiment) is to avoid generation of
executable multiple times for every gem that is 'exefied'.

The executable is generated once (inside a temporary directory) and placed
in 'data' directory relative to the gem installation itself.

This is later used without compilation.

It is pending defer the detection and usage of DevKit to the compile
phase (right now DevKit is added always).